+++ src/mars_patcher/common_types.py
@@ -1,9 +1,12 @@
 from typing import Annotated, TypeAlias
 
-from mars_patcher.auto_generated_types import Areaid, Typeu8
+import mars_patcher.mf.auto_generated_types as types_mf
+import mars_patcher.zm.auto_generated_types as types_zm
 
-AreaId: TypeAlias = Areaid
-RoomId: TypeAlias = Typeu8
+TypeU8: TypeAlias = types_mf.Typeu8 | types_zm.TypeU8
+
+AreaId: TypeAlias = types_mf.Areaid | types_zm.AreaId
+RoomId: TypeAlias = TypeU8
 
 AreaRoomPair = tuple[AreaId, RoomId]