+++ src/mars_patcher/text.py
@@ -2,8 +2,9 @@ import json
 from enum import Enum
 from functools import cache
 
-from mars_patcher.constants.game_data import character_widths, file_screen_text_ptrs
-from mars_patcher.data import get_data_path
+from mars_patcher.constants.game_data import character_widths
+from mars_patcher.mf.constants.game_data import file_screen_text_ptrs
+from mars_patcher.mf.data import get_data_path
 from mars_patcher.rom import Region, Rom
 
 SPACE_CHAR = 0x40
@@ -20,8 +21,12 @@ VALUE_MARKUP_TAG = {
     "STOP_SOUND": (0xA000, 12),
     "WAIT": (0xE100, 8),
 }
+ADAM = 0xE200
+SAMUS = 0xE201
+FEDERATION = 0xE202
 BREAKING_CHARS = {SPACE_CHAR, NEXT, NEWLINE}
 NEWLINE_CHARS = {NEXT, NEWLINE}
+SPEAKER_CHARS = {ADAM, FEDERATION, SAMUS}
 
 KANJI_START = 0x4A0
 KANJI_WIDTH = 10
@@ -112,12 +117,13 @@ def encode_text(
     string: str,
     max_width: int = MAX_LINE_WIDTH,
     centered: bool = False,
-) -> list[int]:
+) -> bytes:
     char_map = get_char_map(rom.region)
     char_widths_addr = character_widths(rom)
     text: list[int] = []
     line_width = 0
     line_number = 0
+    current_speaker = ADAM
 
     prev_break: int | None = None
     width_since_break = 0
@@ -148,9 +154,24 @@ def encode_text(
                         char_val = char_map.get(f"[{tag_str}]")
                         if char_val is None:
                             raise ValueError(f"Invalid markup tag '{tag_str}'")
+
+                    did_speaker_change = False
+                    if char_val in SPEAKER_CHARS and char_val != current_speaker:
+                        did_speaker_change = True
+                        current_speaker = char_val
+
+                    if char_val in NEWLINE_CHARS or did_speaker_change:
+                        prev_break = len(text)
+                        width_since_break = 0
+                        line_width = 0
+                        if char_val == NEXT or did_speaker_change:
+                            line_number = 0
+                        else:
+                            line_number += 1
                     text.append(char_val)
                     markup_tag = None
                 else:
+                    # Still parsing markup tag
                     markup_tag.append(char)
                 continue
         else:
@@ -215,7 +236,12 @@ def encode_text(
         text.append(NEWLINE)
 
     text.append(END)
-    return text
+
+    text_bytes = bytearray()
+    for val in text:
+        text_bytes.append(val & 0xFF)
+        text_bytes.append(val >> 8)
+    return bytes(text_bytes)
 
 
 def write_seed_hash(rom: Rom, seed_hash: str) -> None: